+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/subs.cljs
@@ -290,3 +290,9 @@
         (not credentials-completed?)
         (not env-variables-completed?)
         (not registries-completed?))))
+
+
+(reg-sub
+  ::check-dct
+  (fn [db]
+    (::spec/check-dct db)))