+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/spec.cljs
@@ -52,6 +52,8 @@
 
 (s/def ::error-message any?)
 
+(s/def ::check-dct any?)
+
 
 (s/def ::db (s/keys :req [::deploy-modal-visible?
                           ::loading-deployment?
@@ -77,7 +79,8 @@
                           ::active-step
                           ::data-step-active?
                           ::step-states
-                          ::error-message]))
+                          ::error-message
+                          ::check-dct]))
 
 
 (def step-states {:data           {:step-id :data
@@ -121,4 +124,5 @@
                ::license-accepted?               false
                ::price-accepted?                 false
                ::error-message                   nil
-               ::step-states                     step-states})
+               ::step-states                     step-states
+               ::check-dct                       nil})